requirements.txt
djangorestframework
djangorestframework_jwt
oauth2client
gspread
markdown
django-filter